requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
pymongo>=4.6.0
dnspython>=2.4.2

# Image Processing
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80

# Machine Learning - CPU only versions for compatibility
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.37.0
sentence-transformers>=2.3.0

# Utilities
numpy>=2.0.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
